added double

+++ TempConverter/Program.cs
@@ -6,11 +6,11 @@ namespace TempConverter
     {
         static void Main(string[] args)
         {
-            int celcius, fahrenheit;
+            double celcius, fahrenheit;
             Console.WriteLine("This program will ask the user to enter the temperature in degree celsius and it will convert it to degrees fahrenheit");
             Console.WriteLine("Celcius       Fahrenheit");
             Console.WriteLine("=======       ==========");
-            //
+            
         }
     }
 }
